+++ driver/include/vortex.h
@@ -72,7 +72,7 @@ int vx_upload_kernel_bytes(vx_device_h device, const void* content, size_t size)
 int vx_upload_kernel_file(vx_device_h device, const char* filename);
 
 // get performance counters
-int vx_get_perf(vx_device_h device, uint64_t* cycles, uint64_t* instrs);
+int vx_get_perf(vx_device_h device, size_t* cycles, size_t* instrs);
 
 #ifdef __cplusplus
 }
